feat: add place-card component
