+++ src/project/components/app/app.tsx
@@ -1,0 +1,9 @@
+import MainPage from '../../pages/main-page/main-page';
+
+function App() {
+  return (
+    <MainPage />
+  );
+}
+
+export default App;
