+++ src/project/components/app/place-card/place-card.tsx
